fix: error solving loadStorage function
